refactor: Festival, School, Stage 레거시 코드 제거
