--- backend/src/main/java/com/festago/festival/application/FestivalStageService.java
@@ -1,8 +1,0 @@
-package com.festago.festival.application;
-
-import com.festago.festival.dto.DetailFestivalResponse;
-
-public interface FestivalStageService {
-
-    DetailFestivalResponse findDetail(Long festivalId);
-}
