--- backend/src/main/java/com/festago/festival/repository/FestivalRepositoryCustom.java
@@ -1,10 +1,0 @@
-package com.festago.festival.repository;
-
-import com.festago.festival.domain.Festival;
-import java.time.LocalDate;
-import java.util.List;
-
-public interface FestivalRepositoryCustom {
-
-    List<Festival> findByFilter(FestivalFilter filter, LocalDate currentTime);
-}
