--- backend/src/main/java/com/festago/school/dto/SchoolUpdateRequest.java
@@ -1,14 +1,0 @@
-package com.festago.school.dto;
-
-import jakarta.validation.constraints.NotBlank;
-
-/**
- * @deprecated API 버저닝이 적용되면 해당 클래스 삭제
- */
-@Deprecated(forRemoval = true)
-public record SchoolUpdateRequest(
-    @NotBlank(message = "domain은 공백일 수 없습니다.") String domain,
-    @NotBlank(message = "name은 공백일 수 없습니다.") String name
-) {
-
-}
