+++ backend/src/main/java/com/festago/ticketing/dto/MemberTicketResponse.java
@@ -1,7 +1,6 @@
 package com.festago.ticketing.dto;
 
 import com.festago.stage.domain.Stage;
-import com.festago.stage.dto.StageResponse;
 import com.festago.ticketing.domain.EntryState;
 import com.festago.ticketing.domain.MemberTicket;
 import java.time.LocalDateTime;
@@ -12,8 +11,9 @@ public record MemberTicketResponse(
     LocalDateTime entryTime,
     EntryState state,
     LocalDateTime reservedAt,
-    StageResponse stage,
-    MemberTicketFestivalResponse festival) {
+//    StageResponse stage, // DTO 객체 새롭게 만들 것
+    MemberTicketFestivalResponse festival
+) {
 
     public static MemberTicketResponse from(MemberTicket memberTicket) {
         Stage stage = memberTicket.getStage();
@@ -23,7 +23,6 @@ public record MemberTicketResponse(
             memberTicket.getEntryTime(),
             memberTicket.getEntryState(),
             memberTicket.getCreatedAt(),
-            StageResponse.from(stage),
             MemberTicketFestivalResponse.from(stage.getFestival()));
     }
 }
