+++ backend/src/test/java/com/festago/festival/repository/MemoryFestivalRepository.java
@@ -2,8 +2,6 @@ package com.festago.festival.repository;
 
 import com.festago.festival.domain.Festival;
 import java.lang.reflect.Field;
-import java.time.LocalDate;
-import java.util.List;
 import java.util.Objects;
 import java.util.Optional;
 import java.util.concurrent.ConcurrentHashMap;
@@ -44,15 +42,5 @@ public class MemoryFestivalRepository implements FestivalRepository {
     @Override
     public void deleteById(Long festivalId) {
         memory.remove(festivalId);
-    }
-
-    @Override
-    public void flush() {
-        // NOOP
-    }
-
-    @Override
-    public List<Festival> findByFilter(FestivalFilter filter, LocalDate currentTime) {
-        throw new UnsupportedOperationException();
     }
 }
